Rn73 revert (#4380)
* Revert "Merge pull request #4366 from wix/bugfix/update_slider"

This reverts commit f4c65939f5af6f74aedbcd7d5511403d4ccd5251, reversing
changes made to 546d07f565b399b149ded11f511ee53ccd7cabee.

* Revert "Merge pull request #4369 from wix/bugfix/broken-post-install"

This reverts commit 3083a58de0a2ab3323d1d468666847f7e41960f4, reversing
changes made to b6aaf30f5188755cc779a6ef35a39abe7bac4679.

* Revert "Upgrade/rn73 (#4347)"

This reverts commit bbcb472e39adf7bad4d1d1afe03598c820284db0.

+++ detox/android/detox/src/full/java/com/wix/detox/espresso/action/AdjustSliderToPositionAction.kt
@@ -10,7 +10,7 @@ import com.wix.detox.espresso.common.ReactSliderHelper
 import org.hamcrest.Matcher
 import org.hamcrest.Matchers
 
-class AdjustSliderToPositionAction(private val targetPositionPct: Float) : ViewAction {
+class AdjustSliderToPositionAction(private val targetPositionPct: Double) : ViewAction {
     override fun getDescription() = "adjustSliderToPosition"
     override fun getConstraints(): Matcher<View?>? =
         Matchers.allOf( isDisplayed(), isAssignableFrom(AppCompatSeekBar::class.java) )
